Fixed a bug where a guess with the same letter that's confirmed and then deconfirmed would cause that position's regex character set to become empty.
An example test case is as follows, for the solution word "staff".
taxis ??__?
state +++__
stamp +++__
stash +++__
stall +++__
staff +++++

The bug previously occurred after entering the second guess. Because "state"
confirms the "t" in the second position, the deconfirmation of the second "t"
incorrectly removed the "t" in the second position's regex character set. This
caused the solver to not find any matching words after the second guess.

Test case taken from Ludwig's stream.